fixed imports, reverted and put SolvedMaze in lattice_maze.py
