+++ maze_dataset/generation/generators.py
@@ -4,12 +4,14 @@ from typing import Any, Callable
 import numpy as np
 
 from maze_dataset.constants import CoordArray
-from maze_dataset.maze.lattice_maze import (
-    NEIGHBORS_MASK,
+from maze_dataset.maze import (
     ConnectionList,
     Coord,
     LatticeMaze,
     SolvedMaze,
+)
+from maze_dataset.maze.lattice_maze import (
+    NEIGHBORS_MASK,
     _fill_edges_with_walls,
 )
