+++ maze_dataset/tokenization/tokenizer.py
@@ -8,7 +8,7 @@ from transformers.tokenization_utils import BatchEncoding
 
 from maze_dataset.constants import SPECIAL_TOKENS
 from maze_dataset.dataset.dataset import GPTDatasetConfig
-from maze_dataset.maze.lattice_maze import LatticeMaze
+from maze_dataset.maze import LatticeMaze
 from maze_dataset.plotting.plot_maze import MazePlot
 
 # if TYPE_CHECKING:
