+++ tests/unit/maze_dataset/generation/test_maze_dataset.py
@@ -13,7 +13,7 @@ from maze_dataset.dataset.maze_dataset import (
     MazeDatasetConfig,
     register_maze_filter,
 )
-from maze_dataset.maze.lattice_maze import SolvedMaze
+from maze_dataset.maze import SolvedMaze
 from maze_dataset.utils import bool_array_from_string
